Successfully scraped HTML tables and parsed into the full CSV

+++ backend/data_management/scraper.py
@@ -1,10 +1,16 @@
 import time
 
 import requests
+# Disable insecure request warnings, since the website we are scraping from
+# is missing an SSL  certificate
+import urllib3
 from bs4 import BeautifulSoup
+import os
+
+urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
 
 # Each URL is of the form /rosters.nsf/Annual/<year>-<team_abbreviation>.html
-base_url = 'https://jt-sw.com/football/pro/rosters.nsf/Annual/{}-{}.html'
+base_url = 'https://www.jt-sw.com/football/pro/rosters.nsf/Annual/{}-{}'
 team_abbreviations_to_team_names = {
     # Current team abbreviations as of 2019
     'arz': 'Arizona Cardinals',
@@ -15,7 +21,7 @@ team_abbreviations_to_team_names = {
     'chi': 'Chicago Bears',
     'cin': 'Cincinnati Bengals',
     'cle': 'Cleveland Browns',
-    'dal': 'Cleveland Browns',
+    'dal': 'Dallas Cowboys',
     'den': 'Denver Broncos',
     'det': 'Detroit Lions',
     'gb': 'Green Bay Packers',
@@ -49,7 +55,17 @@ team_abbreviations_to_team_names = {
 # after 1997)
 # It is best to start at 1997 to avoid issues like this
 # The dataset only goes until 2019
-years = [year for year in range(1997, 2019)]
+years = [year for year in range(1997, 2020)]
+
+
+def get_file_path(year, team_abbreviation):
+    # Define the path to the yearly_roster_data relative to this script
+    current_dir = __file__.rstrip('scraper.py')
+    parent_dir = current_dir + '../'
+    roster_dir = parent_dir + 'yearly_roster_data/'
+    file_name = f'{year}-{team_abbreviation}.html'
+    file_path = roster_dir + file_name
+    return file_path
 
 
 def make_request(url):
@@ -67,7 +83,7 @@ def make_request(url):
     # Send a GET request to the URL
     while retry_count > 0:
         # The website we are scraping from has no SSL certificate
-        response = requests.get(url,verify=False)
+        response = requests.get(url, verify=False)
         if response.status_code in (400, 404):  # Indicates a bad request
             raise requests.exceptions.InvalidURL(f'{url} is an unknown URL')
         elif response.status_code == 429:  # Indicates too many requests
@@ -86,32 +102,34 @@ def make_request(url):
 
 
 def download_yearly_roster_data(year, team, file_path):
-    with open(file_path, 'w') as f:
-        # If we successfully created the file, write our html to it
-        # Get the url corresponding to the current year and team
-        url = base_url.format(year, team)
-        # Try to make a GET request to the URL
-        try:
-            response = make_request(url)
-        except requests.exceptions.InvalidURL:
-            # If we are here, we have tried a team abrevation for a
-            # year when that team did not exist, for example the st louis
-            # rams in 2019, after they had changed cities to become Los
-            # Angeles Rams
-            return
-        except (requests.exceptions.TooManyRedirects,
-                requests.exceptions.RequestException) as e:
-            raise RuntimeError(
-                f'Encountered {e} when making a URL request')
-
-        # Strip the web page of unnecessary information, return the
-        # relevant roster table for the given year
-        # Save the html table into a file in our yearly roster data folder
-        # Create the file with our new data
-        roster_html = extract_roster_table_from_page(response.text)
-        print(roster_html)
-        # Write a string representation of the page's HTML for the roster table
-        f.write(str(roster_html))
+    # If we successfully created the file, write our html to it
+    # Get the url corresponding to the current year and team
+    url = base_url.format(year, team)
+    # Try to make a GET request to the URL
+    try:
+        response = make_request(url)
+    except requests.exceptions.InvalidURL:
+        # If we are here, we have tried a team abbreviation for a
+        # year when that team did not exist, for example the st louis
+        # rams in 2019, after they had changed cities to become Los
+        # Angeles Rams
+        return
+    except (requests.exceptions.TooManyRedirects,
+            requests.exceptions.RequestException) as e:
+        raise RuntimeError(
+            f'Encountered {e} when making a URL request')
+    # Strip the web page of unnecessary information, return the relevant
+    # roster table for the given year
+    # Save the html table into a file in our yearly roster data folder
+    # Create the file with our new data
+    # Check if the file already exists to avoid unnecessary downloads
+    if not os.path.exists(file_path):
+        with open(file_path, 'w') as f:
+            roster_html = extract_roster_table_from_page(response.text)
+            print(f"Printing HTML for {year} {team}")
+            # Write a string representation of the page's HTML for the roster
+            # table
+            f.write(str(roster_html))
 
 
 def extract_roster_table_from_page(page):
@@ -123,13 +141,12 @@ def extract_roster_table_from_page(page):
     :return: BeautifulSoup object containing the MVP table.
     """
     soup = BeautifulSoup(page, 'html.parser')
-    soup.find('table').find('table')
     # Extract the specific table containing roster data
-    roster_table = soup.find('table').find('table')
+    roster_table = soup.find_all('table')[-1]
     return roster_table
 
 
-def scrape_yearly_roster_data(years):
+def scrape_yearly_roster_data():
     """
     Scrapes & downloads yearly roster data and saves it into HTML files.
 
@@ -140,12 +157,7 @@ def scrape_yearly_roster_data(years):
     for year in years:
         # For each team abbreivation name:
         for team in team_abbreviations_to_team_names:
-            # Define the path to the yearly_roster_data relative to this script
-            current_dir = __file__.rstrip('scraper.py')
-            parent_dir = current_dir + '../'
-            roster_dir = parent_dir + 'yearly_roster_data/'
-            file_name = f'{year}-{team}.html'
-            file_path = roster_dir + file_name
+            file_path = get_file_path(year, team)
             try:
                 download_yearly_roster_data(year, team, file_path)
             except FileExistsError:
